requirements.txt
pytest == 8.1.1
selene == 2.0.0rc7
allure-pytest == 2.13.2
python-dotenv == 1.0.1
Appium-Python-Client == 3.1.1
pydantic == 2.6.1
requests==2.31.0